FLINK-5911: Extract job display functions to flink_tools (PR 3/4)
Moves get_flink_job_name, should_job_info_be_shown, async job/checkpoint
fetchers, and the full jobs table formatting out of status.py into
flink_tools.py as get_flink_job_name, should_job_info_be_shown,
_fetch_flink_job_details, _fetch_flink_job_checkpoints, and
format_flink_jobs_table. status.py is reduced to API orchestration only.
